user-actor-my-pets - convert ListMyPetsHandler into a handleUserGetMyPets function and move it into lib/handlers
