+++ src/lib/actions/user-get-my-pets.ts
@@ -2,28 +2,25 @@
 
 import APP from "@/lib/app";
 import { getAuthenticatedUserActor } from "@/lib/auth";
-import { ListMyPetsHandler } from "@/lib/handlers/list-my-pets-handler";
-import { ListMyPetsResponse } from "../models/user-models";
+import { MyDog } from "../models/user-models";
 import { Err, Ok, Result } from "../result";
 import { ERR_401_NO_ACTOR, ERR_500_SERVER_ERROR } from "./action-errors";
+import { handleUserGetMyPets } from "../handlers/handle-user-get-my-pets";
 
 export async function userGetMyPets(): Promise<
-  Result<
-    ListMyPetsResponse,
-    typeof ERR_401_NO_ACTOR | typeof ERR_500_SERVER_ERROR
-  >
+  Result<MyDog[], typeof ERR_401_NO_ACTOR | typeof ERR_500_SERVER_ERROR>
 > {
   const actor = await getAuthenticatedUserActor();
   if (actor === null) {
     return Err(ERR_401_NO_ACTOR);
   }
   try {
-    const handler = new ListMyPetsHandler({
+    const myDogs = await handleUserGetMyPets({
+      userId: actor.getUserId(),
       dbPool: await APP.getDbPool(),
       dogMapper: await APP.getDogMapper(),
     });
-    const response = await handler.handle(actor);
-    return Ok(response);
+    return Ok(myDogs);
   } catch {
     return Err(ERR_500_SERVER_ERROR);
   }
