+++ tests/handlers/handle-user-get-my-pets.test.ts
@@ -1,28 +1,25 @@
 import { Pool } from "pg";
-import { withDb } from "./_db_helpers";
+import { withDb } from "../_db_helpers";
 import {
   DogDetails,
-  DogGen,
   DogSecureOii,
   DogSpec,
-  UserRecord,
   YesNoUnknown,
 } from "@/lib/data/db-models";
-import { UserActor } from "@/lib/user/user-actor";
 import {
   getDogDetails,
   getDogMapper,
   getDogOii,
   getDogSecureOii,
-  getUserActorConfig,
   insertDog,
   insertUser,
-} from "./_fixtures";
-import { ListMyPetsHandler } from "@/lib/handlers/list-my-pets-handler";
+} from "../_fixtures";
+import { handleUserGetMyPets } from "@/lib/handlers/handle-user-get-my-pets";
 import { dbInsertDog } from "@/lib/data/db-dogs";
 import { DOG_STATUS } from "@/lib/models/bark-models";
+import { DogMapper } from "@/lib/data/dog-mapper";
 
-describe("ListMyPetsHandler", () => {
+describe("handleUserGetMyPets", () => {
   const USER_IDX = 71;
   const DOG_IDX = 64;
 
@@ -33,10 +30,12 @@ describe("ListMyPetsHandler", () => {
       dogIndices?: number[];
     },
   ): Promise<{
-    handler: ListMyPetsHandler;
-    userRecord: UserRecord;
-    userActor: UserActor;
-    dogGens: DogGen[];
+    args: {
+      userId: string;
+      dbPool: Pool;
+      dogMapper: DogMapper;
+    };
+    userId: string;
   }> {
     const userRecord = await insertUser(userIdx, dbPool);
     const dogGens = await Promise.all(
@@ -44,37 +43,33 @@ describe("ListMyPetsHandler", () => {
         insertDog(dogIdx, userRecord.userId, dbPool),
       ),
     );
-    const userActor = new UserActor(
-      userRecord.userId,
-      getUserActorConfig(dbPool),
-    );
-    const handler = new ListMyPetsHandler({
-      dbPool,
-      dogMapper: getDogMapper(),
-    });
+    const userId = userRecord.userId;
+    const dogMapper = getDogMapper();
     return {
-      handler,
-      userRecord,
-      userActor,
-      dogGens,
+      args: {
+        userId,
+        dbPool,
+        dogMapper,
+      },
+      userId,
     };
   }
 
   it("should return empty list when user has no dogs", async () => {
     await withDb(async (dbPool) => {
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX);
-      const response = await handler.handle(userActor);
-      expect(response.dogs).toEqual([]);
+      const { args } = await getScenario(dbPool, USER_IDX);
+      const dogs = await handleUserGetMyPets(args);
+      expect(dogs).toEqual([]);
     });
   });
   it("should return a list of dogs belonging to the user", async () => {
     await withDb(async (dbPool) => {
       const dogIndices = [2, 3, 5, 7];
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX, {
+      const { args } = await getScenario(dbPool, USER_IDX, {
         dogIndices,
       });
-      const response = await handler.handle(userActor);
-      const receivedName = response.dogs.map((dog) => dog.dogName);
+      const dogs = await handleUserGetMyPets(args);
+      const receivedName = dogs.map((dog) => dog.dogName);
       const expectedNames = await Promise.all(
         dogIndices.map(async (idx) => {
           const oii = await getDogOii(idx);
@@ -86,34 +81,34 @@ describe("ListMyPetsHandler", () => {
   });
   it("should set status to INCOMPLETE when both breed and weight are unspecified", async () => {
     await withDb(async (dbPool) => {
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX);
+      const { args, userId } = await getScenario(dbPool, USER_IDX);
       const details: DogDetails = await getDogDetails(DOG_IDX, {
         dogBreed: "",
         dogWeightKg: null,
       });
       const secureOii: DogSecureOii = await getDogSecureOii(DOG_IDX);
       const spec: DogSpec = { ...details, ...secureOii };
-      const dogGen = await dbInsertDog(dbPool, userActor.getUserId(), spec);
-      const response = await handler.handle(userActor);
-      expect(response.dogs[0].dogStatus).toEqual(DOG_STATUS.INCOMPLETE);
+      const dogGen = await dbInsertDog(dbPool, userId, spec);
+      const dogs = await handleUserGetMyPets(args);
+      expect(dogs[0].dogStatus).toEqual(DOG_STATUS.INCOMPLETE);
     });
   });
   it("should set status to INELIGIBLE when dog weight is under 20KG", async () => {
     await withDb(async (dbPool) => {
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX);
+      const { args, userId } = await getScenario(dbPool, USER_IDX);
       const details: DogDetails = await getDogDetails(DOG_IDX, {
         dogWeightKg: 19.9,
       });
       const secureOii: DogSecureOii = await getDogSecureOii(DOG_IDX);
       const spec: DogSpec = { ...details, ...secureOii };
-      const dogGen = await dbInsertDog(dbPool, userActor.getUserId(), spec);
-      const response = await handler.handle(userActor);
-      expect(response.dogs[0].dogStatus).toEqual(DOG_STATUS.INELIGIBLE);
+      const dogGen = await dbInsertDog(dbPool, userId, spec);
+      const dogs = await handleUserGetMyPets(args);
+      expect(dogs[0].dogStatus).toEqual(DOG_STATUS.INELIGIBLE);
     });
   });
   it("should set status to PERMANENTLY_INELIGIBLE when dog ever pregnant", async () => {
     await withDb(async (dbPool) => {
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX);
+      const { args, userId } = await getScenario(dbPool, USER_IDX);
       const details: DogDetails = await getDogDetails(DOG_IDX, {
         dogEverPregnant: YesNoUnknown.YES,
 
@@ -123,16 +118,14 @@ describe("ListMyPetsHandler", () => {
       });
       const secureOii: DogSecureOii = await getDogSecureOii(DOG_IDX);
       const spec: DogSpec = { ...details, ...secureOii };
-      const dogGen = await dbInsertDog(dbPool, userActor.getUserId(), spec);
-      const response = await handler.handle(userActor);
-      expect(response.dogs[0].dogStatus).toEqual(
-        DOG_STATUS.PERMANENTLY_INELIGIBLE,
-      );
+      const dogGen = await dbInsertDog(dbPool, userId, spec);
+      const dogs = await handleUserGetMyPets(args);
+      expect(dogs[0].dogStatus).toEqual(DOG_STATUS.PERMANENTLY_INELIGIBLE);
     });
   });
   it("should set status to PERMANENTLY_INELIGIBLE when dog ever received blood transfusion", async () => {
     await withDb(async (dbPool) => {
-      const { handler, userActor } = await getScenario(dbPool, USER_IDX);
+      const { args, userId } = await getScenario(dbPool, USER_IDX);
       const details: DogDetails = await getDogDetails(DOG_IDX, {
         dogEverReceivedTransfusion: YesNoUnknown.YES,
 
@@ -142,11 +135,9 @@ describe("ListMyPetsHandler", () => {
       });
       const secureOii: DogSecureOii = await getDogSecureOii(DOG_IDX);
       const spec: DogSpec = { ...details, ...secureOii };
-      const dogGen = await dbInsertDog(dbPool, userActor.getUserId(), spec);
-      const response = await handler.handle(userActor);
-      expect(response.dogs[0].dogStatus).toEqual(
-        DOG_STATUS.PERMANENTLY_INELIGIBLE,
-      );
+      const dogGen = await dbInsertDog(dbPool, userId, spec);
+      const dogs = await handleUserGetMyPets(args);
+      expect(dogs[0].dogStatus).toEqual(DOG_STATUS.PERMANENTLY_INELIGIBLE);
     });
   });
 });
